Require verification before dropping the db

+++ onboarding/management/commands/create_organization.py
@@ -29,11 +29,14 @@ class Command(BaseCommand):
         organization_domain = args[1]
 
         reset = options.get('reset')
-        if reset:
+        verify = raw_input('are you sure you want to reset the db?: ')
+        if reset and verify == 'yes':
             cursor = connection.cursor()
             cursor.execute('DROP SCHEMA PUBLIC CASCADE')
             cursor.execute('CREATE SCHEMA PUBLIC')
             subprocess.call(['python', 'manage.py', 'migrate', '--noinput'])
+        elif reset:
+            raise CommandError('Cancelling create with reset. Failed verification.')
 
         # XXX we shouldn't have an admin token, we should have some way of
         # generating one on the fly though
